+++ src/main/java/org/ytu/hr/core/services/ISignUpService.java
@@ -7,7 +7,7 @@ import java.util.Date;
 
 public interface ISignUpService {
     Integer getCandidateID();
-    Integer getCitizenID();
+    Long getCitizenID();
     Integer getPhoneNumber();
     String getFirstName();
     String getLastName();
